requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.ini
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.12
    # via -r requirements.ini
aiosignal==1.4.0
    # via aiohttp
amqp==5.3.1
    # via kombu
appdirs==1.4.4
    # via zeep
asgiref==3.9.1
    # via django
async-timeout==5.0.1
    # via
    #   aiohttp
    #   redis
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
    #   zeep
billiard==4.2.1
    # via celery
black==22.12.0
    # via -r requirements.ini
cached-property==2.0.1
    # via zeep
celery==5.5.3
    # via
    #   -r requirements.ini
    #   django-celery-beat
    #   flower
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
defusedxml==0.7.1
    # via zeep
dj-database-url==2.3.0
    # via -r requirements.ini
django==4.2.23
    # via
    #   -r requirements.ini
    #   dj-database-url
    #   django-celery-beat
    #   django-ckeditor
    #   django-cors-headers
    #   django-filter
    #   django-jalali
    #   django-js-asset
    #   django-taggit
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   model-bakery
django-celery-beat==2.8.1
    # via -r requirements.ini
django-ckeditor==6.5.1
    # via -r requirements.ini
django-cors-headers==3.14.0
    # via -r requirements.ini
django-filter==24.3
    # via -r requirements.ini
django-jalali==7.3.0
    # via -r requirements.ini
django-js-asset==3.1.2
    # via django-ckeditor
django-taggit==3.1.0
    # via -r requirements.ini
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.16.0
    # via
    #   -r requirements.ini
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via -r requirements.ini
drf-spectacular==0.28.0
    # via -r requirements.ini
et-xmlfile==2.0.0
    # via openpyxl
flower==2.0.1
    # via -r requirements.ini
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gunicorn==23.0.0
    # via -r requirements.ini
humanize==4.12.3
    # via flower
idna==3.10
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
isodate==0.7.2
    # via zeep
jalali-core==1.0.0
    # via jdatetime
jdatetime==5.2.0
    # via django-jalali
jsonschema==4.25.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
kavenegar==1.1.2
    # via -r requirements.ini
kombu==5.5.4
    # via celery
lxml==6.0.0
    # via zeep
model-bakery==1.11.0
    # via -r requirements.ini
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
openpyxl==3.1.5
    # via -r requirements.ini
packaging==25.0
    # via
    #   gunicorn
    #   kombu
pathspec==0.12.1
    # via black
pillow
    # via -r requirements.ini
platformdirs==4.3.8
    # via black
prometheus-client==0.22.1
    # via flower
prompt-toolkit==3.0.51
    # via click-repl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==6.1.1
    # via -r requirements.ini
psycopg[binary]==3.2.5
    # via -r requirements.ini
psycopg-binary==3.2.5
    # via psycopg
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.ini
python-dotenv==1.0.1
    # via -r requirements.ini
pytz==2025.2
    # via
    #   flower
    #   zeep
pyyaml==6.0.2
    # via drf-spectacular
redis==4.6.0
    # via -r requirements.ini
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   kavenegar
    #   requests-toolbelt
    #   zeep
requests-toolbelt==1.0.0
    # via zeep
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   python-dateutil
    #   zeep
sqlparse==0.5.3
    # via django
stem==1.8.1
    # via -r requirements.ini
tomli==2.2.1
    # via black
tornado==6.5.1
    # via flower
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   asgiref
    #   dj-database-url
    #   multidict
    #   psycopg
    #   referencing
tzdata==2025.2
    # via
    #   django-celery-beat
    #   kombu
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
xlrd==1.1.0
    # via -r requirements.ini
yarl==1.20.1
    # via aiohttp
zeep==3.4.0
    # via -r requirements.ini
